+++ .travis.yml
@@ -40,6 +40,8 @@ install:
     fi
   # Compile
   - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && make"
+  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && pip install ."
+
   # the version of testflo is too new, we downgrade to the last version that supports py2
   - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && pip install testflo==1.3.5 "
 
@@ -47,7 +49,7 @@ script:
   # We need to source the mdolab bashrc before running anything
   - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && testflo . -v -n 1"
   # check that new CGNS files have been generated
-  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR/python/examples && ls -l */*.cgns"
+  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR/examples && ls -l */*.cgns"
 
 after_script:
   - docker rm -f app
