--- __init__.py
@@ -1,1 +1,0 @@
-from .python.pyHyp import pyHyp, pyHypMulti
