+++ examples/m6/grid/m6_surf.blk
@@ -1,0 +1,327 @@
+# blocking file written by ICEM CFD - version 14.0
+# written on Tue Apr 23 16:58:08 2013
+
+MATERIALS
+{
+}
+SURFACE_PARAMS 6
+{
+F_78 1e+10 0 0 1
+F_155 1e+10 0 0 1
+F_227 1e+10 0 0 1
+F_299 1e+10 0 0 1
+F_366 1e+10 0 0 1
+F_426 1e+10 0 0 1
+}
+FAMILIES
+{
+PART_1 1
+PART_2 1
+PART_3 1
+PART_4 1
+PART_5 1
+PART_6 1
+SOLID 1
+DEFAULT_EDGE 1
+DEFAULT_VERTEX 1
+}
+SHARED_WALLS
+{
+}
+PREPOINTS 25 internal_ids
+{
+0 1.143599868 0.0005683206837 1.19630003 V_89
+1 0.6906846762 5.80272399e-08 1.19630003 V_85
+2 9.728860277e-07 1.033483272e-07 0 V_82
+3 0.8058999777 0.0005683206837 0 V_80
+4 1.143599868 -0.0005683206837 1.19630003 V_166
+5 0.6906846762 5.80272399e-08 1.19630003 V_162
+6 9.728860277e-07 1.033483272e-07 0 V_159
+7 0.8058999777 -0.0005683206837 0 V_157
+8 0.8058999777 -0.0005683206837 0 V_238
+9 1.143599868 -0.0005683206837 1.19630003 V_234
+10 1.143599868 0.0005683206837 1.19630003 V_231
+11 0.8058999777 0.0005683206837 0 V_229
+12 1.143757105 0 1.196856737 V_306
+13 0.6906890273 5.802695568e-08 1.19630754 V_303
+14 1.143599868 0.0005683206837 1.19630003 V_301
+15 1.143757105 0 1.196856737 V_373
+16 0.6906890273 5.802695568e-08 1.19630754 V_370
+17 1.143599868 -0.0005683206837 1.19630003 V_368
+18 1.143599868 -0.0005683206837 1.19630003 V_433
+19 1.143757105 0 1.196856737 V_430
+20 1.143599868 0.0005683206837 1.19630003 V_428
+21 0.004950443283 -0.01092151366 0 pnt.00
+22 0.6934659481 -0.006137568038 1.19630003 pnt.01
+23 0.004951323383 0.01092176046 0 pnt.02
+24 0.6934664845 0.006137706339 1.19630003 pnt.03
+}
+COMPOSITE_CURVES 22
+{
+0 1 1e+10 0 0 E_93 {
+E_93
+}
+14 1 1e+10 0 0 E_304 {
+E_304
+}
+21 1 1e+10 0 0 crv.02 {
+crv.02
+}
+3 2 1e+10 0 0 E_83 {
+E_83
+E_160
+}
+10 1 1e+10 0 0 E_235 {
+E_235
+}
+17 1 1e+10 0 0 E_371 {
+E_371
+}
+6 1 1e+10 0 0 E_163 {
+E_163
+}
+13 1 1e+10 0 0 E_307 {
+E_307
+}
+20 1 1e+10 0 0 E_431 {
+E_431
+}
+2 1 1e+10 0 0 E_86 {
+E_86
+}
+9 1 1e+10 0 0 E_239 {
+E_239
+}
+16 1 1e+10 0 0 E_374 {
+E_374
+}
+5 1 1e+10 0 0 E_167 {
+E_167
+}
+12 1 1e+10 0 0 E_310 {
+E_310
+}
+19 1 1e+10 0 0 E_434 {
+E_434
+}
+1 1 1e+10 0 0 E_90 {
+E_90
+}
+8 1 1e+10 0 0 E_242 {
+E_242
+}
+15 1 1e+10 0 0 E_377 {
+E_377
+}
+22 1 1e+10 0 0 crv.03 {
+crv.03
+}
+4 1 1e+10 0 0 E_170 {
+E_170
+}
+11 1 1e+10 0 0 E_232 {
+E_232
+}
+18 1 1e+10 0 0 E_437 {
+E_437
+}
+}
+BOUNDARY_FAMILY 0 0 0 0 0 0
+DIMENSION geometry 3 mesh 2 third 3 dim_to_mesh 3
+GRID 0 4
+{
+-0.2287823208 2
+9.728860277e-07 257
+1.143917441 2
+1.372700735 2
+}
+GRID 1 4
+{
+-0.05517371967 2
+-0.03940979764 17
+0.03940981254 2
+0.05517373458 2
+}
+GRID 2 5
+{
+-0.243270421 2
+0 161
+1.18963858 17
+1.216352105 2
+1.459622526 2
+}
+GRID 3 2
+{
+0 2
+1 2
+}
+CONSTRAINTS
+{
+ + 0:1 >= 256
+ + 1:1 >= 16
+ + 2:1 >= 160
+ + 2:2 >= 16
+}
+NODES
+{
+0.004950443283 -0.01092151366 0 0 21 21 x 0 { 1 1 1 }
+0.8058999777 -0.0005683206837 0 0 8 37 x 0 { 2 1 1 }
+0.004951323383 0.01092176046 0 0 23 25 x 0 { 1 2 1 }
+0.8058999777 0.0005683206837 0 0 11 41 x 0 { 2 2 1 }
+0.678344809 -0.006242632974 1.170026887 1 22 43 x 0 { 1 1 2 } - 22:0.978038
+1.136183369 -0.00056832068 1.170026887 1 9 44 x 0 { 2 1 2 } - 9:0.021962
+0.678345322 0.006242773716 1.170026905 1 21 45 x 0 { 1 2 2 } - 21:0.978038
+1.136183369 0.00056832068 1.170026887 1 0 46 x 0 { 2 2 2 } - 0:0.021962
+0.6934659481 -0.006137568038 1.19630003 0 22 22 x 0 { 1 1 3 }
+1.143599868 -0.0005683206837 1.19630003 0 18 38 x 0 { 2 1 3 }
+0.6934664845 0.006137706339 1.19630003 0 24 26 x 0 { 1 2 3 }
+1.143599868 0.0005683206837 1.19630003 0 20 42 x 0 { 2 2 3 }
+}
+SUPERBLOCKS
+{
+{ 10 2 2 2 -1 0:1:0 1:1:0 3:0:0 { 21
+ 37 25 41 21 37 25 41 } 3 0 0 }
+{ 1 2 2 2 7 2:1:0 1:1:0 3:0:0 { 21
+ 43 25 45 21 43 25 45 } 1 1 0 }
+{ 2 2 2 2 7 0:1:0 2:1:0 3:0:0 { 21
+ 37 43 44 21 37 43 44 } 2 1 0 }
+{ 3 2 2 2 7 1:1:0 2:1:0 3:0:0 { 37
+ 41 44 46 37 41 44 46 } 4 1 0 }
+{ 4 2 2 2 7 2:1:0 0:1:0 3:0:0 { 25
+ 45 41 46 25 45 41 46 } 5 1 0 }
+{ 5 2 2 2 7 0:1:0 1:1:0 3:0:0 { 22
+ 38 26 42 22 38 26 42 } 6 1 0 }
+{ 7 2 2 2 7 1:1:0 2:2:1 3:0:0 { 22
+ 26 43 45 22 26 43 45 } 1 1 0 }
+{ 6 2 2 2 7 2:2:1 0:1:0 3:0:0 { 22
+ 43 38 44 22 43 38 44 } 2 1 0 }
+{ 8 2 2 2 7 2:2:1 1:1:0 3:0:0 { 38
+ 44 42 46 38 44 42 46 } 4 1 0 }
+{ 9 2 2 2 7 0:1:0 2:2:1 3:0:0 { 26
+ 42 45 46 26 42 45 46 } 5 1 0 }
+}
+SUPERFACES
+{
+}
+EDGE_CURVES no_interpolate node_nums
+{
+21 37 0
+{
+1 3 0 1
+}
+21 43 0
+{
+1 22 0 1
+}
+21 25 0
+{
+1 3 0 1
+}
+37 41 0
+{
+1 8 0 1
+}
+37 44 0
+{
+1 9 0 1
+}
+25 45 0
+{
+1 21 0 1
+}
+25 41 0
+{
+1 3 0 1
+}
+41 46 0
+{
+1 0 0 1
+}
+43 22 0
+{
+1 22 0 1
+}
+43 44 1 -1 -1 0 s
+{
+2 0 0 1
+0.921125293 -0.0205883561 1.19921923 2 0 0 1
+}
+44 38 0
+{
+1 9 0 1
+}
+45 46 1 -1 -1 0 s
+{
+2 0 0 1
+0.9211252928 0.02058835607 1.199219232 2 0 0 1
+}
+45 26 0
+{
+1 21 0 1
+}
+46 42 0
+{
+1 0 0 1
+}
+22 38 3
+{
+2 0 0 1
+0.739252631 -0.0088936073 1.2092229 2 0 0 1
+0.85293407 -0.0142586231 1.21250951 2 0 0 1
+0.984027912 -0.0105275575 1.20983456 2 0 0 1
+}
+22 26 7
+{
+2 0 0 1
+0.6929377241 -0.005114201523 1.196836516 2 0 0 1
+0.6924568332 -0.003686967567 1.197451567 2 0 0 1
+0.6921186365 -0.001936393932 1.197921424 2 0 0 1
+0.6920154084 4.347199343e-08 1.198080922 2 0 0 1
+0.6921187321 0.001936497727 1.197921454 2 0 0 1
+0.6924568609 0.003687255628 1.197451825 2 0 0 1
+0.6929383019 0.005114264843 1.196836467 2 0 0 1
+}
+38 42 7
+{
+2 0 0 1
+1.143739967 -0.0005154760303 1.196429522 2 0 0 1
+1.143868737 -0.0003716337871 1.196477117 2 0 0 1
+1.143936019 -0.0001904164882 1.196522531 2 0 0 1
+1.143956751 5.141465583e-06 1.196547915 2 0 0 1
+1.143933825 0.0002019042267 1.196537048 2 0 0 1
+1.143862623 0.0003829127067 1.196497781 2 0 0 1
+1.143730115 0.0005203562217 1.196440613 2 0 0 1
+}
+26 42 3
+{
+2 0 0 1
+0.7392526307 0.00889360733 1.209222896 2 0 0 1
+0.8529340702 0.01425862306 1.212509513 2 0 0 1
+0.9840279115 0.0105275575 1.209834562 2 0 0 1
+}
+}
+DEFAULT_BUNCHING_LAW default 2
+EDGE_RATIOS node_nums
+{
+21 37 { linked 21 25 } 2 { linked 41 37 } 2 1 2 1e+10
+21 43 0.005 2 { linked 43 22 } 2 1 2 5.5063e+09
+21 25 8.1664e+09 2 8.1664e+09 2 0 0 8.16635e+09
+37 41 5.2842e+08 2 5.2842e+08 2 0 0 5.2842e+08
+37 44 0.005 2 { linked 44 38 } 2 1 2 4.95881e+09
+25 45 0.005 2 { linked 45 26 } 2 1 2 5.5063e+09
+25 41 { linked 21 25 } 2 { linked 41 37 } 2 1 2 9.99999e+09
+41 46 0.005 2 { linked 46 42 } 2 1 2 4.95881e+09
+43 22 0 2 0.000555205 2 0 2 5.5063e+09
+43 44 { linked 43 45 } 2 { linked 44 46 } 2 1 2 0
+43 45 0 2 0 2 0 0 0
+44 38 0 2 0.00025 2 0 2 4.95881e+09
+44 46 0 2 0 2 0 0 0
+45 46 { linked 43 45 } 2 { linked 44 46 } 2 1 2 0
+45 26 0 2 0.000555205 2 0 2 5.5063e+09
+46 42 0 2 0.00025 2 0 2 4.95881e+09
+22 38 { linked 21 25 } 2 { linked 41 37 } 2 1 2 5.59163e+09
+22 26 5e+09 2 5e+09 2 0 0 5e+09
+38 42 4.24948e+08 2 4.24948e+08 2 0 0 4.24947e+08
+26 42 { linked 21 25 } 2 { linked 41 37 } 2 1 2 5.59163e+09
+}
+return
