+++ pyhyp/__init__.py
@@ -1,0 +1,3 @@
+__version__ = '2.1.0'
+
+from .pyHyp import pyHyp, pyHypMulti
