--- python/__init__.py
@@ -1,1 +1,0 @@
-# This file is empty
