--- python/examples/717/mdo_tutorial_ffd.fmt
@@ -1,74 +1,0 @@
-            1
-            6           2           6
-  -1.1766643020000001E-002  0.9938716787839999        1.999510000588000     
-    3.005148322392000        4.010786644196000        5.016424966000000     
-  -1.1766643020000001E-002  0.9938716787839999        1.999510000588000     
-    3.005148322392000        4.010786644196000        5.016424966000000     
-    1.492714954784000        2.358371375147200        3.224027795510400     
-    4.089684215873600        4.955340636236801        5.820997056600000     
-    1.492714954784000        2.358371375147200        3.224027795510400     
-    4.089684215873600        4.955340636236801        5.820997056600000     
-    2.997196552588000        3.722871071510400        4.448545590432801     
-    5.174220109355200        5.899894628277600        6.625569147200000     
-    2.997196552587999        3.722871071510400        4.448545590432800     
-    5.174220109355200        5.899894628277600        6.625569147200000     
-    4.501678150391999        5.087370767873599        5.673063385355200     
-    6.258756002836800        6.844448620318400        7.430141237800000     
-    4.501678150391999        5.087370767873599        5.673063385355199     
-    6.258756002836799        6.844448620318399        7.430141237799999     
-    6.006159748196001        6.451870464236801        6.897581180277601     
-    7.343291896318401        7.789002612359200        8.234713328400000     
-    6.006159748196001        6.451870464236800        6.897581180277600     
-    7.343291896318400        7.789002612359199        8.234713328400000     
-    7.510641346000000        7.816370160600000        8.122098975200000     
-    8.427827789800000        8.733556604400000        9.039285419000000     
-    7.510641346000000        7.816370160600000        8.122098975200000     
-    8.427827789800000        8.733556604400000        9.039285419000000     
-  -0.3160394621000000      -0.3160394621000000      -0.3160394621000000     
-  -0.3160394621000000      -0.3160394621000000      -0.3160394621000000     
-   0.3344310415000000       0.3344310376800000       0.3344310338600000     
-   0.3344310300400000       0.3344310262200000       0.3344310224000000     
-  -0.2731421910600000      -0.2731421909760001      -0.2731421908920000     
-  -0.2731421908080001      -0.2731421907240001      -0.2731421906400000     
-   0.2886978048800001       0.2886978018240001       0.2886977987680000     
-   0.2886977957120001       0.2886977926560000       0.2886977896000000     
-  -0.2302449200200000      -0.2302449198520000      -0.2302449196840000     
-  -0.2302449195160000      -0.2302449193480000      -0.2302449191800000     
-   0.2429645682600000       0.2429645659680000       0.2429645636760000     
-   0.2429645613840000       0.2429645590920000       0.2429645568000000     
-  -0.1873476489800000      -0.1873476487280000      -0.1873476484760000     
-  -0.1873476482240000      -0.1873476479720000      -0.1873476477200000     
-   0.1972313316400000       0.1972313301120001       0.1972313285840000     
-   0.1972313270560000       0.1972313255280000       0.1972313240000000     
-  -0.1444503779400000      -0.1444503776040000      -0.1444503772679999     
-  -0.1444503769320000      -0.1444503765960000      -0.1444503762600000     
-   0.1514980950200000       0.1514980942560000       0.1514980934920000     
-   0.1514980927280000       0.1514980919640000       0.1514980912000000     
-  -0.1015531069000000      -0.1015531064800000      -0.1015531060600000     
-  -0.1015531056400000      -0.1015531052200000      -0.1015531048000000     
-   0.1057648584000000       0.1057648584000000       0.1057648584000000     
-   0.1057648584000000       0.1057648584000000       0.1057648584000000     
-  -1.0000000000000000E-003 -1.0000000000000000E-003 -1.0000000000000000E-003
-  -1.0000000000000000E-003 -1.0000000000000000E-003 -1.0000000000000000E-003
-  -1.0000000000000000E-003 -1.0000000000000000E-003 -1.0000000000000000E-003
-  -1.0000000000000000E-003 -1.0000000000000000E-003 -1.0000000000000000E-003
-    2.809917728000000        2.809917728000000        2.809917728000000     
-    2.809917728000000        2.809917728000000        2.809917728000000     
-    2.809917728000000        2.809917728000000        2.809917727999999     
-    2.809917728000000        2.809917728000000        2.809917728000000     
-    5.620835456000000        5.620835456000001        5.620835456000000     
-    5.620835456000001        5.620835456000002        5.620835456000002     
-    5.620835455999999        5.620835456000001        5.620835455999998     
-    5.620835456000000        5.620835456000000        5.620835456000000     
-    8.431753184000000        8.431753184000000        8.431753184000000     
-    8.431753184000000        8.431753184000000        8.431753184000000     
-    8.431753184000000        8.431753184000000        8.431753183999998     
-    8.431753183999998        8.431753183999998        8.431753183999998     
-    11.24267091200000        11.24267091200000        11.24267091200000     
-    11.24267091200000        11.24267091200000        11.24267091200000     
-    11.24267091200000        11.24267091200000        11.24267091200000     
-    11.24267091200000        11.24267091200000        11.24267091200000     
-    14.05358864000000        14.05358864000000        14.05358864000000     
-    14.05358864000000        14.05358864000000        14.05358864000000     
-    14.05358864000000        14.05358864000000        14.05358864000000     
-    14.05358864000000        14.05358864000000        14.05358864000000
